Renaming for clarity
